+++ src/NeatMapper/Options/Configuration/Internal/CustomMapsConfiguration.cs
@@ -4,7 +4,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Runtime.CompilerServices;
-using System.Security.AccessControl;
 using System.Threading.Tasks;
 
 namespace NeatMapper {