+++ tests/NeatMapper.Tests/Mapping/MergeMapsTests.cs
@@ -455,7 +455,7 @@ namespace NeatMapper.Tests.Mapping {
 
 			MappingOptionsUtils.context = null;
 			Assert.AreEqual("4", _mapper.Map(2, ""));
-			Assert.IsNull(MappingOptionsUtils.context.MappingOptions.GetOptions<NestedMappingContext>());
+			Assert.IsFalse(MappingOptionsUtils.context.MappingOptions.HasOptions<NestedMappingContext>());
 
 			Assert.AreEqual("-6", _mapper.Map(-3, ""));
 			Assert.AreEqual("0", _mapper.Map(0, ""));
@@ -678,11 +678,11 @@ namespace NeatMapper.Tests.Mapping {
 				Assert.AreEqual(2, result.Categories.First());
 				Assert.AreEqual(3, result.Categories.Last());
 
-				Assert.IsNull(Maps.productOptions.GetOptions<NestedMappingContext>());
+				Assert.IsFalse(Maps.productOptions.HasOptions<NestedMappingContext>());
 				// Should use same context for nested maps
 				Assert.AreEqual(2, Maps.categoryOptions.Count);
 				Assert.AreEqual(1, Maps.categoryOptions.Distinct().Count());
-				Assert.IsTrue(Maps.categoryOptions.All(o => o.GetOptions<NestedMappingContext>() != null));
+				Assert.IsTrue(Maps.categoryOptions.All(o => o.HasOptions<NestedMappingContext>()));
 
 
 				// Factory
@@ -709,11 +709,11 @@ namespace NeatMapper.Tests.Mapping {
 				Assert.AreEqual(2, result2.Categories.First());
 				Assert.AreEqual(3, result2.Categories.Last());
 
-				Assert.IsNull(Maps.productOptions.GetOptions<NestedMappingContext>());
+				Assert.IsFalse(Maps.productOptions.HasOptions<NestedMappingContext>());
 				// Should use same context for nested maps
 				Assert.AreEqual(2, Maps.categoryOptions.Count);
 				Assert.AreEqual(1, Maps.categoryOptions.Distinct().Count());
-				Assert.IsNotNull(Maps.categoryOptions.First().GetOptions<NestedMappingContext>());
+				Assert.IsTrue(Maps.categoryOptions.First().HasOptions<NestedMappingContext>());
 			}
 
 			{
@@ -881,7 +881,7 @@ namespace NeatMapper.Tests.Mapping {
 		[TestMethod]
 		public void ShouldCheckCanMapWithAdditionalMaps() {
 			var options = new CustomMergeAdditionalMapsOptions();
-			options.AddMap<string, int>((s, d, _) => s?.Length ?? 0, c => c.MappingOptions.GetOptions<ProjectionCompilationContext>() == null);
+			options.AddMap<string, int>((s, d, _) => s?.Length ?? 0, c => !c.MappingOptions.HasOptions<ProjectionCompilationContext>());
 			var mapper = new CustomMapper(null, null, options);
 
 			Assert.IsTrue(mapper.CanMapMerge<string, int>());